requirements.txt
torch
torchvision
transformers
datasets
pillow
scikit-learn
opendataval
pandas
matplotlib